ChatHandlerSendMessageHttp implemented
