+++ .env.example
@@ -26,7 +26,8 @@ CACHE_BACKEND=redis
 CACHE_SERVICE_PREFIX=ai_bot:
 CACHE_TTL_SECONDS=300
 
-REDIS_HOST=localhost # local | cache (docker)
+# local | cache (docker)
+REDIS_HOST=localhost
 REDIS_PORT=6379
 REDIS_DB=0
 REDIS_MAX_CONNECTIONS=50
