+++ README.md
@@ -44,6 +44,13 @@ there are 2 simple chat clients for testings
 * [test_multi_client.html](./tests/e2e/test_multi_client.html) **_(recommended)_**
 * [test_client.html](./tests/e2e/test_client.html)
 
+### Send Chat message via HTTP
+
+```bash
+ curl -X POST http://127.0.0.1:20003/chat/http/bot/send-message/session-001 -H "Content-Type: application/json" -d '{"message": "Hello World"}'; echo
+```
+
+
 ### Install black (formatter) globally
 
 ```bash
